wasm-gc: extract wasip2 call-site emits into builtins_wasip2.rs
`body/builtins.rs` had reached 2351 lines, of which the last ~610
were the wasip2-specific `emit_*_wasip2` call-site lowerings
(`Console.print/error/warn/readLine`, `Args.get`, `Env.get`,
`Time.{unixMs,now,sleep}`, `Random.{int,float}`, `Disk.{exists,
readText}`). Each of those is the wasip2 counterpart of an
aver-bridge handler that stays in the main switch — same shape
as the orchestrator/helper split landed in commit bca661b4 for
`module.rs` and `wasip2_helpers.rs`.

Move them into a new sibling `body/builtins_wasip2.rs`. The dispatch
in `emit_dotted_builtin` now `use`s the moved functions through
`super::builtins_wasip2::{...}`. Pure code movement plus
`pub(super)` on the relocated fns; no logic change.

  src/codegen/wasm_gc/body/
    builtins.rs            2351 → 1740  (-611)
    builtins_wasip2.rs     new   630

All wasip2 e2e fixtures and unit / POC tests still green.
